+++ backend/db/client.py
@@ -1,6 +1,9 @@
 import os
 
 from supabase import Client, create_client
+from dotenv import load_dotenv
+
+load_dotenv()
 
 # Initialization of Supabase client goes here
 url: str = os.environ.get("SUPABASE_URL")